Make sure the axes array will be integer-typed.

+++ theano/tensor/basic.py
@@ -1360,7 +1360,8 @@ class MaxAndArgmax(Op):
                                  dtype=node.outputs[0].dtype)
         # Numpy does not support multiple axes for argmax
         # Work around
-        keep_axes = numpy.array([i for i in range(x.ndim) if i not in axes])
+        keep_axes = numpy.array([i for i in range(x.ndim) if i not in axes],
+                                dtype='int64')
         # Not-reduced axes in front
         transposed_x = numpy.transpose(x, numpy.concatenate((keep_axes, axes)))
         reshaped_x = transposed_x.reshape(transposed_x.shape[:len(keep_axes)] +
